refactor: share provider quota cards
