+++ .agents/SESSIONS/2026-06-30.md
@@ -1,0 +1,72 @@
+# Sessions: 2026-06-30
+
+**Summary:** Exhausted quota cards now use the blocking reset presentation across dashboard sections
+
+---
+
+## Session 1: Align exhausted weekly reset cards
+
+**Duration:** ~30 minutes
+**Status:** Complete
+
+### What was done
+
+- Verified `origin/master` matches the current checkout at `026740c`, so the stale-app suspicion was not caused by this worktree being behind master.
+- Checked open PRs and found only #63, which is scoped to multiple Claude Code accounts and compact popover behavior rather than the dashboard reset-card mismatch.
+- Added shared exhausted-limit state to dashboard provider snapshots.
+- Updated Overview and Limits provider cards to show the same blocking reset panel used by the menu bar card when any quota window is exhausted.
+- Updated Costs provider breakdown cards to show the matching provider reset panel when the provider is currently blocked by an exhausted quota.
+
+### Files changed
+
+- `MeterBar/Views/UsageDashboardView.swift` - Reuses `BlockingLimitResetCounter` across Overview, Limits, and Costs provider cards.
+
+### Decisions
+
+- **Decision:** Reuse the existing `BlockingLimitResetCounter` instead of creating a new dashboard-only card.
+  - **Rationale:** It keeps the wording and countdown selection identical to the menu bar card, including weekly reset text such as `Weekly reset` and `in 2d 13h`.
+- **Decision:** For Costs, match provider cost rows to the visible provider snapshot by service and prefer an exhausted snapshot when multiple accounts share a service.
+  - **Rationale:** Cost data is service-level, while quota data can be account-level for Claude, so the safest visible signal is the blocked account if one exists.
+
+### Verification
+
+- `git diff --check` passed.
+- `swift build` passed.
+- `swift test --filter ResetCountdownTests` was blocked by this environment missing XCTest under the active Command Line Tools developer directory.
+- `xcodebuild -project MeterBar.xcodeproj -scheme MeterBar -configuration Debug -destination 'platform=macOS' CODE_SIGNING_ALLOWED=NO build` was blocked because full Xcode is not selected/installed (`xcodebuild` requires Xcode, active directory is `/Library/Developer/CommandLineTools`).
+
+---
+
+## Session 2: Consolidate provider quota card components
+
+**Duration:** ~45 minutes
+**Status:** Complete
+
+### What was done
+
+- Extracted shared provider quota UI into `ProviderQuotaViews.swift`.
+- Replaced the duplicate popover/dashboard snapshot and limit models with `ProviderQuotaSnapshot` and `ProviderQuotaLimit`.
+- Replaced the duplicate popover/dashboard provider cards and limit rows with `ProviderQuotaCard` and `ProviderQuotaLimitRow`.
+- Moved shared provider primitives out of `MenuBarView.swift`: provider logo kind, logo view/cache, provider title, reset countdown views, blocking reset counter, extra-usage pill, and usage bar.
+- Updated the menu bar popover, Overview cards, Limits cards, and Costs blocked-provider reset panel to use the same shared quota state.
+- Preserved the three visual densities through `ProviderQuotaCardVariant`: popover, dashboard overview, and dashboard limits.
+
+### Files changed
+
+- `MeterBar/Views/ProviderQuotaViews.swift` - New shared quota model/view layer.
+- `MeterBar/Views/MenuBarView.swift` - Now builds shared provider snapshots and renders shared quota cards.
+- `MeterBar/Views/UsageDashboardView.swift` - Now renders shared quota cards in Overview/Limits and uses shared snapshots for Costs reset state.
+
+### Decisions
+
+- **Decision:** Use one provider quota snapshot model across popover and dashboard.
+  - **Rationale:** The reset-window selection, percent-left calculation, tertiary limit title, exhausted state, and provider ids should not drift by surface.
+- **Decision:** Keep surface-specific visual density as enum variants instead of separate components.
+  - **Rationale:** The popover and dashboard need different typography/spacing, but they should share the same quota behavior and row rendering.
+
+### Verification
+
+- `git diff --check` passed.
+- Duplicate type scan confirmed the old popover/dashboard quota type names are gone.
+- `swift build` passed.
+- `swift test --filter ResetCountdownTests` is still blocked by the local XCTest availability issue (`no such module 'XCTest'`).
